add adjust button on movement row and and show adjust quantity form
