+++ src/components/activities/pharmacyActivity/pharmaceuticalStock/components/stockActions/StockActions.tsx
@@ -1,8 +1,6 @@
 import ExitToAppIcon from "@mui/icons-material/ExitToApp";
 import LogoutIcon from "@mui/icons-material/Logout";
 import Button from "components/accessories/button/Button";
-import { AdjustQuantityForm } from "components/activities/pharmacyActivity/pharmaceuticalStock/components/forms/adjustQuantityForm/AdjustQuantityForm";
-import StockModal from "components/activities/pharmacyActivity/pharmaceuticalStock/components/modal/Modal";
 import { MovementDTO } from "generated";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
@@ -16,34 +14,6 @@ export interface StockActionsHandle {
 export const StockActions = React.forwardRef<StockActionsHandle, {}>(
   (_, ref) => {
     const { t } = useTranslation();
-
-    const [openStockModal, setOpenStockModal] = useState<boolean>(false);
-    const [selectedMovement, setSelectedMovement] = useState<
-      MovementDTO | undefined
-    >();
-
-    const handleAdjustClick = (movement: MovementDTO) => {
-      setSelectedMovement(movement);
-      setOpenStockModal(true);
-    };
-
-    const handleCloseModal = () => {
-      setOpenStockModal(false);
-      setSelectedMovement(undefined);
-    };
-
-    const handleAdjustSubmit = (values: {
-      movement: MovementDTO;
-      newQuantity: number;
-    }) => {
-      console.log("Adjust quantity:", values);
-      // TODO: Call API to update movement quantity
-      handleCloseModal();
-    };
-
-    React.useImperativeHandle(ref, () => ({
-      handleAdjustClick,
-    }));
 
     return (
       <div className="pharmaceuticalStock__actions">
@@ -64,16 +34,6 @@ export const StockActions = React.forwardRef<StockActionsHandle, {}>(
           {t("pharmacy.stock.stockReport")}
         </Button>
         <div className="separator"></div>
-        <Button
-          className="discharge_button"
-          type="button"
-          variant="contained"
-          color="inherit"
-          startIcon={<LogoutIcon sx={{ transform: "rotate(90deg)" }} />}
-          onClick={() => setOpenStockModal(true)}
-        >
-          Adjust Quantity
-        </Button>
         <Link to={"./discharge-movement"}>
           <Button
             type="button"
@@ -95,14 +55,6 @@ export const StockActions = React.forwardRef<StockActionsHandle, {}>(
             {t("pharmacy.stock.charge")}
           </Button>
         </Link>
-
-        <StockModal open={openStockModal} onClose={handleCloseModal}>
-          <AdjustQuantityForm
-            movement={selectedMovement}
-            onSubmit={handleAdjustSubmit}
-            onCancel={handleCloseModal}
-          />
-        </StockModal>
       </div>
     );
   }
